+++ tests/data/configs/web.php
@@ -1,4 +1,17 @@
 <?php
+
+if (!defined('DB_DSN')) {
+    define('DB_DSN', 'foo'); // presevers errors
+}
+
+if (!defined('DB_USER')) {
+    define('DB_USER', 'foo'); // presevers errors
+}
+
+if (!defined('DB_PASS')) {
+    define('DB_PASS', 'foo'); // presevers errors
+}
+
 return [
     'id' => 'testenv',
     'siteTitle' => 'Luya Tests',